requirements.txt
# Core dependencies
google-adk
python-dotenv>=1.0.0

# Database
supabase>=0.7.1
neo4j>=5.8.1
gqlalchemy>=1.2.0  # For Memgraph

# Bluesky API
atproto>=0.0.21
aiohttp>=3.8.4  # For async HTTP requests

# NLP and Sentiment Analysis
nltk>=3.8.1
vaderSentiment>=3.3.2
scikit-learn>=1.2.2  # For basic topic modeling

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.2
tqdm>=4.65.0  # For progress bars

# Async
asyncio>=3.4.3

# Logging
structlog>=23.1.0

# Development
pytest>=7.3.1
pytest-cov>=4.1.0
